% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_weekly_timeslots.R
\name{geom_weekly_timeslots}
\alias{geom_weekly_timeslots}
\title{Construct geom_weekly_timeslots function}
\usage{
geom_weekly_timeslots(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{(\code{SlotsDetails}) A \link{SlotsDetails} data table.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
(\code{ggplot}) A ggplot object.
}
\description{
Construct geom_weekly_timeslots function
}
\examples{
geom_weekly_timeslots(data = SlotsDetails())

}
