% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlotsDetails.R
\name{SlotsDetails}
\alias{SlotsDetails}
\title{Construct SlotsDetails value object}
\usage{
SlotsDetails(
  start_datetime = NA,
  end_datetime = NA,
  title = NA_character_,
  body = list(NA_character_)
)
}
\arguments{
\item{start_datetime}{(\code{POSIXct}) Is the start of the time slot.For example "2022-03-16 10:00:00".}

\item{end_datetime}{(\code{POSIXct}) Is the end of the time slot.For example "2022-03-16 11:30:00".}

\item{title}{(\code{character}) Is the title of the time slot. For example "News 1".}

\item{body}{(\code{list}) Is a nested list with name value pairs of the time slot details. For example list(cost = "$1200", TARP = 5).}
}
\value{
(\code{data.frame}) Is a long dataframe of time slots.
}
\description{
Construct SlotsDetails value object
}
\examples{
SlotsDetails()

}
