% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.slots-package.R
\docType{package}
\name{time.slots-package}
\alias{time.slots}
\alias{time.slots-package}
\title{time.slots: Package to Create Weekly Heatmaps}
\description{
A collection of functions and scripts to generate weekly timetables which can be overlaid with a heatmap to show trends.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.org/annalectnz/time.slots}
  \item Report bugs at \url{https://bitbucket.org/annalectnz/time.slots/issues}
}

}
\author{
\strong{Maintainer}: Mukti Rathod \email{mukti.rathod@annalect.com}

Other contributors:
\itemize{
  \item Harel Lustiger \email{harel.lustiger@annalect.com} [contributor]
}

}
\keyword{internal}
